# docs/swagger.yaml
components:
  schemas:
...
    github_com_leseb_openresponses-gw_pkg_core_schema.CreateConversationRequest:
      properties:
        metadata:
          type: object
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.ListConnectorsResponse:
      properties:
        data:
          description: Array of connectors
          items:
            $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Connector'
          type: array
          uniqueItems: false
        first_id:
          description: ID of first item
          type: string
        has_more:
          description: Whether there are more results
          type: boolean
        last_id:
          description: ID of last item
          type: string
        object:
          description: Always "list"
          type: string
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.RegisterConnectorRequest:
      properties:
        connector_id:
          description: Required
          type: string
        connector_type:
          description: Required, must be "mcp"
          type: string
        metadata:
          type: object
        server_label:
          type: string
        url:
          description: Required
          type: string
      type: object
...
info:
  contact:
    name: Open Responses Gateway
    url: https://github.com/leseb/openresponses-gw
  description: |-
    100% Open Responses Specification Compliant Gateway.
    Based on: https://github.com/openresponses/openresponses

    This gateway provides:
    - **Core API**: Full Open Responses spec compliance (POST /v1/responses)
    - **Extended APIs**: Conversations, Prompts, Files, Vector Stores, Connectors
    - **Dual Mode**: Standalone HTTP server or Envoy ExtProc integration

    Streaming: All 24 event types from Open Responses spec
    Request Echo: All request parameters returned in response
    Multimodal: Support for text, images, files, video
  title: Open Responses Gateway API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/connectors:
    get:
      parameters:
      - description: Cursor for pagination
        in: query
        name: after
        schema:
          type: string
      - description: Cursor for pagination (backwards)
        in: query
        name: before
        schema:
          type: string
      - description: Number of items (1-100, default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Sort order: asc or desc (default desc)'
        in: query
        name: order
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListConnectorsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: List connectors
      tags:
      - Connectors
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.RegisterConnectorRequest'
                description: Register connector request
                summary: request
        description: Register connector request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Connector'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: Register connector
      tags:
      - Connectors
  /v1/connectors/{connector_id}:
    delete:
      parameters:
      - description: Connector ID
        in: path
        name: connector_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.DeleteConnectorResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Delete connector
      tags:
      - Connectors
    get:
      parameters:
      - description: Connector ID
        in: path
        name: connector_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Connector'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Get connector
      tags:
      - Connectors
  /v1/conversations:
    get:
      parameters:
      - description: Cursor for pagination
        in: query
        name: after
        schema:
          type: string
      - description: Cursor for pagination (backwards)
        in: query
        name: before
        schema:
          type: string
      - description: Number of items (1-100, default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Sort order: asc or desc (default desc)'
        in: query
        name: order
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListConversationsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: List conversations
      tags:
      - Conversations
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.CreateConversationRequest'
                description: Create conversation request
                summary: request
        description: Create conversation request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Conversation'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: Create conversation
      tags:
      - Conversations
  /v1/conversations/{id}:
    delete:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.DeleteConversationResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Delete conversation
      tags:
      - Conversations
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Conversation'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Get conversation
      tags:
      - Conversations
  /v1/conversations/{id}/items:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Cursor for pagination
        in: query
        name: after
        schema:
          type: string
      - description: Cursor for pagination (backwards)
        in: query
        name: before
        schema:
          type: string
      - description: Number of items (1-100, default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Sort order: asc or desc (default desc)'
        in: query
        name: order
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListConversationItemsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: List conversation items
      tags:
      - Conversations
    post:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.AddConversationItemsRequest'
                description: Items to add (max 20)
                summary: request
        description: Items to add (max 20)
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.AddConversationItemsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Add conversation items
      tags:
      - Conversations
  /v1/files:
    get:
      parameters:
      - description: Cursor for pagination
        in: query
        name: after
        schema:
          type: string
      - description: Cursor for pagination (backwards)
        in: query
        name: before
        schema:
          type: string
      - description: Number of items (1-100, default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Sort order: asc or desc (default desc)'
        in: query
        name: order
        schema:
          type: string
      - description: Filter by purpose
        in: query
        name: purpose
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListFilesResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: List files
      tags:
      - Files
    post:
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              oneOf:
              - title: file
                type: file
              - title: purpose
                type: string
          multipart/form-data:
            schema:
              type: object
        description: 'File to upload | Purpose: assistants, vision, batch, or fine-tune'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.File'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: Upload file
      tags:
      - Files
  /v1/files/{id}:
    delete:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.DeleteFileResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Delete file
      tags:
      - Files
    get:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.File'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Get file
      tags:
      - Files
  /v1/files/{id}/content:
    get:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                type: file
          description: OK
        "400":
          content:
            application/octet-stream:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/octet-stream:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
        "500":
          content:
            application/octet-stream:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: Get file content
      tags:
      - Files
...
  /v1/prompts/{id}/versions:
    get:
      parameters:
      - description: Prompt ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListPromptsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: List prompt versions
      tags:
      - Prompts
...
servers:
- description: Local development server
  url: http://localhost:8080
...